feat: :art: refactor sync task and add task for remote storage provider upload
