+++ storage-node/src/services/sync/tasks/ISyncTask.ts
@@ -1,0 +1,14 @@
+/**
+ * Defines syncronization task abstraction.
+ */
+export interface SyncTask {
+  /**
+   * Returns human-friendly task description.
+   */
+  description(): string
+
+  /**
+   * Performs the task.
+   */
+  execute(): Promise<void>
+}
