+++ storage-node/src/services/sync/tasks/index.ts
@@ -1,0 +1,3 @@
+export * from './DownloadTask'
+export * from './StorageProviderSyncTask'
+export * from './DeleteLocalFileTask'
